foodlab-backend/src/core/security.py
import uuid
from datetime import datetime, timedelta
from typing import Any, Union, Optional
from passlib.context import CryptContext

from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from jose import JWTError, jwt

from src.core.config import settings
from src.utils.unitofwork import IUnitOfWork, UnitOfWork
from src.models.user import User, UserRole

security = HTTPBearer()

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")


async def get_current_user(
        credentials: HTTPAuthorizationCredentials = Depends(security),
        uow: IUnitOfWork = Depends(UnitOfWork)
) -> User:
    """Получить текущего пользователя из токена"""
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )

    try:
        payload = jwt.decode(
            credentials.credentials, settings.SECRET_KEY, algorithms=[settings.ALGORITHM]
        )
        user_id: str = payload.get("sub")
        if user_id is None:
            raise credentials_exception
    except JWTError:
        raise credentials_exception

    async with uow:
        user = await uow.users.get(uuid.UUID(user_id))
        if user is None or not user.is_active:
            raise credentials_exception
        return user


async def require_admin(current_user: User = Depends(get_current_user)) -> User:
    """Требовать роль ADMIN"""
    if current_user.role != UserRole.ADMIN:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="Admin access required"
        )
    return current_user


async def require_canteen(current_user: User = Depends(get_current_user)) -> User:
    """Требовать роль CANTEEN"""
    if current_user.role != UserRole.CANTEEN:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="Canteen access required"
        )
    return current_user


async def require_staff(current_user: User = Depends(get_current_user)) -> User:
    """Требовать роль ADMIN или CANTEEN (для обратной совместимости)"""
    if current_user.role not in [UserRole.ADMIN, UserRole.CANTEEN]:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="Staff access required"
        )
    return current_user

def create_access_token(subject: Union[str, Any], expires_delta: timedelta = None) -> str:
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)

    to_encode = {"exp": expire, "sub": str(subject)}
    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
    return encoded_jwt


def create_refresh_token(subject: Union[str, Any]) -> str:
    expire = datetime.utcnow() + timedelta(days=settings.REFRESH_TOKEN_EXPIRE_DAYS)
    to_encode = {"exp": expire, "sub": str(subject)}
    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
    return encoded_jwt


def verify_password(plain_password: str, hashed_password: str) -> bool:
    return pwd_context.verify(plain_password, hashed_password)


def get_password_hash(password: str) -> str:
    return pwd_context.hash(password)


def verify_token(token: str) -> Optional[str]:
    try:
        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])
        user_id: str = payload.get("sub")
        if user_id is None:
            return None
        return user_id
    except JWTError:
        return None