foodlab-backend/requirements.txt
# Core FastAPI & ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
psycopg2==2.9.11

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-decouple==3.8
pyjwt==2.8.0
cryptography==41.0.7

# Pydantic & Validation
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Payments
stripe==7.7.0

# CORS
python-multipart==0.0.6

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# Environment & Config
python-dotenv==1.0.0

# Async utilities
asgiref==3.7.2
aiofiles==23.2.1

# HTTP Client
requests==2.31.0

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# UUID
uuid7==0.1.0

# Optional: Redis for caching/rate limiting
redis==5.0.1
hiredis==2.2.3

# Optional: Email sending
aiosmtplib==3.0.1

# Optional: Monitoring
sentry-sdk[fastapi]==1.34.0

# Optional: Task queue (Celery)
celery==5.3.4
redis==5.0.1

# Optional: WebSockets
websockets==12.0